Refactor methods between HadoopFsHelper and AvroFsHelper
