+++ gobblin-core/src/main/java/gobblin/source/extractor/hadoop/HadoopFsHelper.java
@@ -13,6 +13,7 @@
 package gobblin.source.extractor.hadoop;
 
 import java.io.IOException;
+import java.io.InputStream;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
@@ -22,24 +23,32 @@ import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.io.compress.CompressionCodec;
+import org.apache.hadoop.io.compress.CompressionCodecFactory;
+
 import com.google.common.base.Strings;
 import gobblin.configuration.ConfigurationKeys;
 import gobblin.configuration.State;
 import gobblin.source.extractor.filebased.FileBasedHelper;
 import gobblin.source.extractor.filebased.FileBasedHelperException;
 import gobblin.source.extractor.filebased.TimestampAwareFileBasedHelper;
+import gobblin.util.HadoopUtils;
 import gobblin.util.ProxiedFileSystemWrapper;
 
 
 /**
  * A common helper that extends {@link FileBasedHelper} and provides access to a files via a {@link FileSystem}.
  */
-public abstract class HadoopFsHelper implements TimestampAwareFileBasedHelper {
+public class HadoopFsHelper implements TimestampAwareFileBasedHelper {
   private final State state;
   private final Configuration configuration;
   private FileSystem fs;
 
-  protected HadoopFsHelper(State state, Configuration configuration) {
+  public HadoopFsHelper(State state) {
+    this(state, HadoopUtils.newConfiguration());
+  }
+
+  public HadoopFsHelper(State state, Configuration configuration) {
     this.state = state;
     this.configuration = configuration;
   }
@@ -119,4 +128,49 @@ public abstract class HadoopFsHelper implements TimestampAwareFileBasedHelper {
           .format("Failed to get last modified time for file at path %s due to error %s", filePath, e.getMessage()), e);
     }
   }
+
+  @Override
+  public long getFileSize(String filePath) throws FileBasedHelperException {
+    try {
+      return this.getFileSystem().getFileStatus(new Path(filePath)).getLen();
+    } catch (IOException e) {
+      throw new FileBasedHelperException(
+          String.format("Failed to get size for file at path %s due to error %s", filePath, e.getMessage()), e);
+    }
+  }
+
+  /**
+   * Returns an {@link InputStream} to the specified file.
+   * <p>
+   * Note: It is the caller's responsibility to close the returned {@link InputStream}.
+   * </p>
+   *
+   * @param path The path to the file to open.
+   * @return An {@link InputStream} for the specified file.
+   * @throws FileBasedHelperException if there is a problem opening the {@link InputStream} for the specified file.
+   */
+  @Override
+  public InputStream getFileStream(String path) throws FileBasedHelperException {
+    try {
+      Path p = new Path(path);
+      InputStream in = this.getFileSystem().open(p);
+      // Account for compressed files (e.g. gzip).
+      // https://github.com/apache/spark/blob/master/core/src/main/scala/org/apache/spark/input/WholeTextFileRecordReader.scala
+      CompressionCodecFactory factory = new CompressionCodecFactory(this.getFileSystem().getConf());
+      CompressionCodec codec = factory.getCodec(p);
+      return (codec == null) ? in : codec.createInputStream(in);
+    } catch (IOException e) {
+      throw new FileBasedHelperException("Cannot open file " + path + " due to " + e.getMessage(), e);
+    }
+  }
+
+  @Override
+  public void close() throws FileBasedHelperException {
+    try {
+      this.getFileSystem().close();
+    } catch (IOException e) {
+      throw new FileBasedHelperException("Failed to close filesystem", e);
+    }
+  }
+
 }
